Introduce make variable for global CFLAGS build.
The following make variable was introduced to the build
configuration file:

BUILD_CFLAGS

That is used to specify the generic CFLAGS to be used for the
BUILD process.  Subordinate Makefiles should use this variable to
initialize the CFLAGS variable that will be used for the actual
compilation.

Use the new infrastructure to define the CAP_TRUST pre-processor
define that is used to set the value that will be used to drop
the CAP_TRUST capability in the workload launch processes for the
trust orchestrators.

Drop the defines from the quixote.h file that had been used when
the old system call interface had been used to manage the TSEM
implementation.