Use pkg-config to locate Xen include files.
In another hard coded dependency that had escaped scrutiny the
build of the SanchoXen implementation was failing secondary to an
inability to locate the xenstore.h include file.

The location of the Xen includes had been hardcoded in the XENSRC
make variable.  This variable was changed to XEN_INCLUDE and the
pkg-config utility was used to set the value of this variable.

Since the output of the pkg-config utility pre-pendends the -I
command-line option to the path specification, the sed utility was
used to strip the option.   This is needed since the path
directive is used with the -isystem command-line option for
embedded compilations.