Rationalize source code clean process.
Previously, a command to globally clean the source code to a
pristine state was inconsistent.  In particular the Sancho
heirarchy was using 'make clobber' to do this.

Update the entire source tree to use 'make distclean' as the
command to do this.

An anomaly was noted with the local build configuration getting
generated, if it did not exist, if a second distclean was run
after the source code was cleaned.  This appears to be due to the
fact the '-include Build.mk' directive in the top level Makefile
was electing to build the Build.mk file to try and fulfill the
include dependency.  This will need to be addressed in future
updates to the build system.