Update HurdLib submodule.
The upgrade to version 9 of the GCC compiler demonstrated various
correctness errors and compiler warnings.  A number of these
were inherent to the HurdLib implementation.

Upgrades were made to HurdLib to correct this issue and add some
functionality improvements.  Upgrade the implementation used by
Quixote to pick up these fixes and changes.