Introduce alternate event parsing infrastructure.
Testing of the SanchoMCU implementations resulted in
non-deterministic hangs where the TMA implementation became
unresponsive.  Extensive testing indicated that these hangs were
being caused by the TMA implementation failing in the regexec()
function.

Some consideration had been previously given to whether or not
regular expression based parsing was appropriate for security
event processing.  Given the difficulty of tracing the regular
expression code in a micro-controller environment the decision
was made forward to proceed forward with a simpler parsing
infrastructure.

In order to support this the following object was introduced:

EventParser

This object encapsulates the functionality needed to extract a
field from a security event description and then extract specific
parameters from this description.

This initial commit introduces the object itself and updates the
build infrastructure to reference the new object.

The initial target for the new parsing infrastructure was the
security event description field and process ID for the event.