Convert COE to new parser infrastructure.
Following the implementation of the new parser infrastructure the
COE object was converted to use this parser.

The COE object has to address the fact that the capability mask
of a process is currently a 64 bit value.  Unfortuantely the
EventParser->get_integer method was designed to work with 32 bit
values.

In order to make the parser more generic the method was converted
to extract a 64 numeric value with the strtoll() function.  The
helper functions in each parsing object are now expected to
test the returned value with respect to the size of the numeric
value that is to be returned.

The SecurityEvent object was converted to address this altered
API.