Avoid false detection of a field description.
The EventParser->extract_field method would misdetect a field
description of the text of the description was a subset of
another string in the event description line.

For example the following field description:

mmap_file{type=0, reqprot=5, prot=5, flags=1048578}

Would get matched by a request for a 'file' directive, ie:

file{flags=32800, uid=0, gid=0, ... }

This issue was fixed by requiring that the event field descriptor
was immediately followed by an opening brace ({) and preceeded by
a space if the event field descriptor was not at the beginning of
the line.