Implement option to block logging of security event descriptions.
The NR52840-DK SanchoMCU implementation was found to be failing
in what appeared to be a non-deterministic fashion.  An extensive
round of debugging was finally able to find the root cause of the
problem which was the depletion of allocatable memory.

The problem was extremely difficult to find given that it does
not appear the standard library implementation that comes with
the ARM GCC compiler does not seem to be returning a NULL pointer
when the memory arena is depleted.  A proper return of a NULL
pointer would have caused the Buffer object to trap the error but
instead the allocator appeared to silently fail.

Given the limited memory budget available on the micro-controller
implementations the decision was made to have the TSEM object
provide the object to not log either standard execution
trajectory or forensics events.  Instead only the security event
points are maintained in MCU memory, which is sufficient for the
SanchoMCU implementation to properly model the security event
descriptions that are passed to it.

The following method was implemented to disable logging of the
security event descriptions:

TSEM->disable_logging

With security event description logging disabled the NRF52840
implementation increased the size of the model that could be
handled from approximately 89 events to at least 1100.