Disable security event description logging for Nordic SanchoMCU.
Testing indicated that the Nordic SanchoMCU implementation
appeared to be failing in a non-deterministic fashion.  The
failure was manifest by a modeling session failing at various and
random points in an execution trajectory.

Additional testing indicated that when a failure occurred it
always occurred after the same number of event descriptions in a
given execution trajectory were processed.  This suggested that
while the problem was non-deterministic for different execution
trajectories it was always deterministic for a given trajectory.

This situation required extensive testing but was ultimately
tracked to depletion of allocatable memory.  Discovering the
source of the problem was complicated by the fact that it appears
as if the standard library implementation provided with the ARM
and MCU implementations was not returning a NULL pointer when
memory allocation failed.

As a result it was difficult to pinpoint the exact cause of the
failure.  What happened instead was that the event field
description would just silently discontinue accumulating the full
field description into the String object.  The result was the
failure to parse the event that in turn caused the SanchoMCU
implementation to lock up.

The problem was fixed by disabling event logging in the TSEM
object instance that was being used in the MCU implementation.

When fed the boot trajectory the NRF52840-DK implementation was
hanging after 89 events were processed.  Disabling the event
logging enabled the full root trajectory of 1,068 events to
be processed.

With only security state point logging enabled the theoretical
limit for the number of events that can be tracked is over 9,000,
with the exact number determined by other consumption of
allocatable memory.

The problem with this approach is that the execution trajectory
is not available for the creation of security maps.  The plan
will be to add tracking of the trajectory in the trust
orchestrator rather than the MCU implementation proper.