Rationalize TSEM control infrastructure handling.
Segmentation faults were noted when the SanchoSGX implementation
was tested with the new TSEM based control infrastructure.

This was traced to the fact that a rather ad-hoc setup was being
to initialize the TSEMcontrol object that was being used to set
the trust status on modeled processes.  The TSEMcontrol object
was not being initialized when the standard quixote-sgx
application was being used which generatedthe segmentation fault.

The strategy employed to address this was to make the TSEMcontrol
object used to discipline processes be contained the state
information for the SanchoSGX object.  The ecall1 (update)
interface was modified to pass this object into the enclave so
that it would be available to the OCALL that is used to set the
trust status for the object.

This allowed the single static instance of the object to be
removed from the quixote-sgx trust orchestrator.  A model that in
retrospect was probably incorrect.