Support output of security trajectory and maps for SanchoSGX.
Both of the SanchoSGX based trust orchestrators were modified to
support the output of either a security map or a security
trajectory when the modeled workload is terminated.  This is
support that was previously implemented in the SanchoUS and
SanchoKernel implementations.

Properly support an execution trajectory required increasing the
size of the buffer that was used to marshall output from the
ECALL that returns the security event descriptions.  The previous
buffer value of 512 was too small to handle the longer security
event descriptions that are being generated by the LSM based TSEM
implementation.