Support SanchoMCU model output.
The MCU based trust orchestrator was the remaining orchestrator
that was not supporting the ability to output a security model
upon termination of a workload.  This commit provides this
support.

The -t command-line option to specify that a security execution
trajectory be output is included but currently generates an empty
file.  This is secondary to the fact that logging of the
execution trajectory descriptions has been disabled secondary to
memory resource issues.