Expand buffer size for event descriptions.
Testing of the quixote-xen trust orchestrator indicated that it
was hanging when the first event was processed.

This was traced to the fact that the buffer that is used to read
from the /sys/fs/tsem/ExternalTMA/NN pseudo-file was sized at 512
bytes.  This is insufficient to hold the length of event
descriptions that have expanded beyond 512 bytes.

Increase the buffer size by 256 bytes to 768 bytes which is a
value longer than any messages currently being generated.