Truncate model/trajectory output files.
A persistent issue that was noted in testing was the fact that if
a model output file had a previous model in it an updated model,
that was less than the size of the previous model, would end up
being corrupted.

This was secondary to the fact that output of the new model would
begin at file offset zero, leaving erroneous model input, from
the previous model.

Address this issue by calling truncate() on the output file with
an offset of zero prior to opening the file.  This causes the
file size to be reset to zero and prevents leftover data in the
file from corrupting the new model.