Call truncate unconditionally.
Testing indicated that the trust orchestrators were generating
error messages when output of a security map was requested to a
file that did not exist.  This was secondary to the fact that the
output of truncate() was being tested for an error return, the
absence of the file generated such an error.

For the time being simply call truncate() unconditionally since
it avoids the generation of an error and detection of an actual
error is arguably not useful.