Implement global definition of loader flags.
In order to continue the simplification of the build process the
BUILD_LDFLAGS was introduced.  This variable is designed to be
the generic definition for the LDFLAGS variable.  Individual
Makefiles can use the BUILD_LDFLAGS variable as the basis for the
actual LDFLAGS for the sub-system binaries.

The Config.mk file adds the '-static' flag to the BUILD_LDFLAGS
if the STATIC variable is set to any value.

The Sancho/Makefile that drives the compilation of the Trusted
Modeling Agents was modified to source the definitions from
global build configuration.

The main build sub-systems can now all be built directly their
top level directories.  Support for building directly from any of
the subordinate directories will be done at a later time.