Set installation path in global build configuration.
The BUILD_INSTPATH variable was added to the global build
configuration in order to establish a default installatino
location.  The BUILD_INSTPATH variable respects the DESTDIR
variable for providing an alternate root for the installation
path.

In each subordinate Makefile the INSTPATH variable is used to
specify the installation location.  Update all of the Makefiles
to set the INSTPATH value based on the BUILD_INSTPATH variable.