Update compilation directives for QuixoteSGX.
Pre-processor defines are used to change the location for the
SanchoSGX enclave and tokens.  The build of the Quixote utilities
was not including the pre-processor define that changes the
location files for a production or release build.  Fix this by
passing the correct pre-processor directive to the compilation.

Also modify the locations from their previous defualt for ESD
based utilitys to the Quixote root.

Since the tooling release that will support the TSEM release will
not be signed by a production key alter the pre-processor defines
so the enclave is loaded in debug mode.