Update build system.
An issue was noted where the compilation environment was not
being fully modified when the GCC compiler was selected.  This
was secondary to the fact that the Config.mk file is querying the
value of the CC variable in order to customize the build.
Unfortunately the build model is that the Build.mk file, that is
included after the Config.mk file is processed is used to alter
the CC variable.

In order to address this the Setup.mk file was introduced that is
designed to configure the build based on the sum value of all all
the variables that have been defined by the Config.mk and
Build.mk files.  This file tests the CC variable and modifies the
appropriate variables for the build.

Since the Build.mk file is fungible the notion of a source
independent configuration file was added.  This file, named
.config, in the top of the build directory is designed to hold
customization values that are to be included every time the
Build.mk file is updated.