Address bash dependency and CFLAGS issues on Ubuntu.
Another bash shell dependency was noted in the compilation of the
MINI-OS version of the pciutils library that supports the
SanchoXen installation.  The command that was copying the support
files into place were using brace delimiters to specify the
basenames of the files to be copied.  This caused execution of
the cp command to fail under the standard POSIX shell.  Fix this
by copying the files individually.

The SanchoXen compilation used the 'pkg-config xenstore' utility
to set the 'system' include path.  Under Ubuntu the pkg-config
command returns nothing since the files are in the standard
/usr/include heirarchy.

This issue was addressed by testing the value of the XEN_INCLUDE
variable that was set with the pkg-config utility.  If the
variable is null set it to /usr/include as a 'last' resort.