Update SanchoXen compilation.
A functionality issue was noted when the SanchoXen TMA was
compiled under Ubuntu 22.04.  A segmentation fault would be
generated whenever the stubdomain implementation was started.

After a somewhat lengthy investigation this was isolated to what
appears to be an issue with version 11 of the GCC compiler that
is supported by that distribution.  This compiler requires a
series of options that disable stack protection and a number of
other features that seem cause issues with the launch of the
domain.

Include these options in the PLATFORM make variable that is
included in all the code that is compiled for the SanchoXen
implementation.

In addition the mini-os checkout was simply using the current
HEAD rather than a verson of the code that was intended for a
particular Xen release.

The following variable was introduced to the global build
configuration:

BUILD_XEN_VERSION

To set the major release version of Xen that the compilation is
being done against.

These changes resulted in the generation of a functional
stubdomain implementation of the TMA.