Implement purge target.
In order to prevent multiple downloads of the build components,
the purge target was added to the build configuration to remove
any downloaded files or cloned repositories.  The purge command
invokes the 'distclean' target from the top level Makefile so
that the source tree is returned to a pristine state.

Invoking the 'distclean' target now results in a source tree that
will be completed recompiled but will not require a subsequent
download of remote components.