Download and deploy Nordic SDK.
The build of the Nordic NRF SanchoMCU implementations is
dependent on a Software Development Kit that Nordic supplies and
is freely available.

Integrate the download and deployment of the SDK in the Support
directory so it can be referenced by the Sancho builds.

The Makefiles for the build of the micro-controller
implementations were modified to reference the global build
configuration to find the location of the development kit.