Drop the config target.
Experience with the new build system indicated that the config
target in the top level Makefile was not useful.

With the introduction of the .config variable to hold local
configuration changes a better approach is to make the Build.mk
file dependent on the .config file.  This will cause the Build.mk
file to be rebuilt if any changes are made to the .config file.