Implement creating /var directory framework.
The creation of the framework needed in the /var heirarchy was
added to the install-bin target.

The following variable was created in the global build
configuration to set the default root for the framework:

BUILD_VARPATH

This variable is constructed with the DESTDIR variable to
accomodate re-direction into an alternate installation heirarchy.