Change root of SGX SDK.
The Quixote SGX development environment use the libraries and
include files from the Intel SGX SDK as well as the pre-compiled
enclaves.

These were formerly rooted in the following location:

/opt/intel

In order to avoid collisions with the Intel libraries the root
for the Quixote specific implementations were moved to the
following location:

/opt/Quixote/share/intel

The build system was modified to use this new location.