Make kernel version consistent with other global configurations.
The specification of the kernel version that the trust
orchestrators are to be compiled against was done with the
following global make variable:

KERNEL_VERSION

This is now inconsistent with the model of using the BUILD_
prefix on all global build variables.

Modify the global kernel configuration to use the following
variable:

BUILD_KERNEL_VERSION

Modify the subordinate Makefiles to set the KERNEL_VERSION
variable based on the new global build variable.