Upgrade SRDE to handle compatibility with current Intel SDK/PSW.
The compilation of the Quixote SGX SRDE was upgraded to be
compatible with the most current release of the Intel SGX System
Development Kit and Platform SoftWare which is version 2.18
plus 16 commits.

The updated trusted runtime environment now returns an error when
the enclave de-initialization ECALL is issued.  The
SRDEenclave->boot_slot method was modified to not report an error
when this ECALL is issued when the destructor for the object is
called.