Add infrastructure to flash NRF52840-DONGLE firmware.
Since the SanchoMCU implementation based on the NRF52840-DONGLE
micro-controller is the most likely device to be flashed at the
level of this release implement infrastructure that simplifies
the process.

In order to adapt to a wide variety of implementations the
following global build variable was added:

BUILD_NRFUTIL

This variable needs to be set with the arguement that is needed
to execute the Python3 based nrfutil utility.

A wrapper script named nrf-flash was created in the
implementation directory.  This script conducts the various steps
needed, using the nrfutil utility, to package the hex encoded
firmware implementation for flashing into a device.