Convert control plane to securityfs.
A major review comment from the first submission of the patch
series was from Greg Kroah-Hartmann who objected to the use of a
dedicated control plane filesystem at /sys/fs/tsem.

The TSEM LSM was modified to move the root of the control plane
to the following location:

/sys/kernel/security/tsem

That is within a filesystem managed by the securityfs filesystem
implementation.

His concern was that the control plane was being advertised as a
sysfs implementation which he indicated was incorrect.  He also
objected to the notion of replicating the filesystem
implementation code that was needed to support an independent
filesystem approach.

This commit modifies the trust orchestrators to use the new
filessytem locations to interrogate an internal trusted modeling
agent implementation and to control the TSEM implementation.

The total reduction of the kernel based filesystem implementation
was 1,105 bytes.