Implement JSON based TSEM event parsing.
The initial review of the TSEM patches, that had been submitted
for mainstream review, brought criticism that a custom format was
being used for output of the security execution trajectories.
Greg Kroah-Hartman recommended that a format such as JSON be
considered.

I questioned Paul Moore, the Linux security sub-system
maintainer, if JSON was acceptable but received no reply to that
query.

Given that Greg Kroah-Hartman, who is considered the second in
command in Linux kernel development, had suggested JSON I elected
to move forward with that format.

The following two objects were implemented in order to support
this new format:

TSEMparser

TSEMevent

The TSEMparser object implements enough of a JSON parser to
implement parsing of the exported security event descriptions.
It is by no means a complete JSON parser and it is dependent on
the form of JSON that is output by the TSEM event export
mechanism.

The TSEMevent object was implemented in order to encapsulate the
functionality that is needed to interpret a TSEM event
description.  After some reflection the decision was made to
implement the following method:

TSEMevent->encode_event

That converts a security event description from JSON format to
the previous format that had been used.  This allows the existing
security event description parsing infrastructure to be used
without modification.

The previously used encoding format was extremely similar to JSON
encoding but more efficient for the given purpose.  An event
description in 'Quixote' formatting is approximately 26% smaller
than the JSON encoding of the event.