Convert quixote-us to JSON based event output.
With the TSEMparser and TSEMevent objects committed and with the
TSEM LSM now outputting JSON encoded security event descriptions
the quixote-us utility was converted to use the TSEMevent object
to interpret a security event description.

As was noted in the documentation for the introduction of the
JSON parsing the TSEMevent->encode_event method is used to encode
a security event description into the previous 'Quixote' format
of encoding before injecting the security event description into
the TMA implementation.

In the case of an export of the system aggregate the TSEMevent
object is used to extract the value of the aggregate and use the
string representation of the object as it had previously used the
string from the exported event.

Stale code that had been previously used to handle a measurement
event was removed as it has remained unused and commented out for
an extended period of time.