Implement generation of randomized TMA authentication key.
The following method was added to the TSEMcontrol object:

TSEMcontrol->generate_key

This method generates a random 256 bit key that will be used to
authenticate processes being modeled in a security modeling
domain.  This value is passed as an argument to the external
namespace creation control command, the value is also appended to
the trusted and untrusted control commands issued by the
->release and ->discipline methods.

The userspace trust orchestrator was updated to generate the
authentication key.  The other trust orchestrators need similar
updates.