Output quixote encoded trajectories.
With the conversion to the output of security event descriptions
in JSON the trust orchestrator for internally modeled domains was
outputing the execution trajectories encoded in JSON.  In
contrast the trajectories for externally modeled domains were
being output in quixote encoded format.

This behavior is secondary to the fact that in the case of an
internally modeled domain the trust orchestrator is reading the
execution trajectories from the kernel pseudo-files.  In the case
of an externally modeled domain the formatting of the events is
done by the trusted modeling agent.

In order to address this the following method was added to the
TSEMevent object:

TSEMevent->extract_event

Since the output from the kernel pseudo-files is not encoded
with an 'export' directive this method is needed in order to
allow the TSEMevent->encode_event to function properly.

The internal trust orchestrator was updated to use the TSEMevent
object to convert entries read from the kernel pseudo-files to
Quixote encoding for subsequent output to the quixote-console
utility.