Handle JSON encoded audit events.
The userspace trust orchestrator was not properly handling JSON
encoded log (audit) events.  This commit implements support for
properly handling these events and fixes a long standing bug in
the TSEM model object.

The following method was added to the TSEMevent object in order
to convert a JSON encoded log event to the previously used
Quixote encoding:

TSEMevent->encode_log

The process_event() function in the userspace orchestrator was
modified to convert the JSON encoded entry into Quixote encoding
before injecting the log event into the model.

In order to further clarify code the following function:

add_TSEM_event()

Was renamed to the following in the trust orchestrator:

add_log()

The TSEM object was simply inserting the String object that was
passed to it into the GAGGLE that holds the list of events.  This
has been a long standing issue that can subsequently cause a
regression when the GAGGLE is referenced or deleted.

The TSEM->add_TSEM_EVENT method was modified to create a String
object which the incoming string is copied into.  At that point
the destructor for the TSEM object becomes responsible for
disposing of the String object.