Properly encode mmap_file with embedded file description.
The kernel TSEM driver was updated to embed a file{} description
in the mmap_file description.  This commit updates the event
parser to handle this change.

The TSEMevent->encode_event method was updated to split the
encoding of the event description into three separate cases:

- standard event

- file_open

- mmap_file

The handling of an mmap_file event was modified to handle the
case of an anonymous mapping separately from a file backed
mapping.  An anonymous mapping is handled like a standard event
while a file backed mapping outputs the mmap description
parameters, followed by the encoding of the file description as a
quixote file{} based field.

In order to handle the embedded file{} description the
_copy_field() function was introduced to handle the copying of an
event description while translating JSON to quixote encoding.
The _copy_field() function was modified to detect an embedded
structure definition and discontinue the copy when that was
detected.

After the field copy is completed the file{} field is extracted
and then translated into a quixote encoded field.

This new architecture required a change to the _find_key() helper
function in order to detect both a standard key/value pair and a
key/structure key value.  The function now searches for the
following string:

"name":_space_

If this sub-string is detected the next string is verified to
check whether it is a " or a {.  A failure to detect that results
in a field field detection failure.

If either initiating character is found that character is added
to the key_value field in order to complete the key description.