Handle new naming nomenclature for effective capability mask.
The Linux TSEM driver was modified in a manner that changed the
key value of the effective capability from 'cap' to 'capeff'.

Modify the COE object in order to properly parse and format the
key/value pair.

In order to make the code more consistent the name of the
characteristic was changed from 'capability' to 'capeff'.