Implement support for user namespace reference.
The Linux TSEM implementation was modified to allow the user
namespace that was used as the basis for uid/gid mappings to be
specified at the time that a modeling domain/namespace was
created.

The 'internal' and 'external' namespace creation verbs were
modified to accept as the first argument one of the following two
values:

init

current

Specifing init causes the initial user namespace to be referenced
for the uig/gid mappings.  This is identical to the previous
behavior.

Specifying an argument of 'current', causes the namespace to use
the current user namespace of the process that is generating the
security event to be referenced for mapping.

The trust orchestrators were modified to use the '-u' argument
as a request that the 'current' namespace behavior be selected.

The TSEMcontrol->external and TSEMcontrol->internal methods were
replaced by the following method:

TSEMcontrol->create_ns

In order to consolidate namespace creation functionality into one
function.  The following enumeration type was surfaced from the
TSEMcontrol object in order to allow selection of the namespace
type and user namespace reference values:

TSEMcontrol_ns_config

The enumeration values specified by this enumeration type are
passed as arguments to the TSEMcontrol->create_ns method in order
to configure the creation of the namespace.