Reset the event description before setting it.
A regression was noted where the same quixote event encoding was
being returned despite different JSON event descriptions having
been set in the object.

This was secondary to the fact that the String object that holds
the event description was not be reset.  Explicitly reset the
String object so that it only holds one event description at a
time.