Output internal TMA model and trajectory in quixote encoding.
The trust orchestrator that implements the in-kernel Trusted
Modeling Agent was generating its execution trajectory in JSON
encoding since that was the native output of the in-kernel TMA.

Introduce the output_trajectory() function to to output the
trajectory in quixote rather than JSON encoding.

With this change all of the modeling infrastructure now operates
in quixote encoding.