Implement support for modeling domain digest configuration.
With the advent of the ability of the TSEM drive to accept a
namespace control message that can specify the hash function that
will be used for security event characteristics compression the
ability to specify the hash function was added to the trust
orchestrators.

The TSEMcontrol->create_ns method was modified to accept an
additional argument, a character string that points to the name
of the hash function that is to be specified for the modeling
domain.  The hash definition used are the hash identification
strings that are used in the Linux shash implementation.

The various hash function specifications that are available can
be found in the 'crypto' directory in the top level of the Linux
kernel sources.  Examples include sha256, sha1, sha512, sha3-224,
sha3-512, sm3 etc.

Specifying a NULL value for the pointer causes the sha256 hash
function to be selected.  The TSEM driver 'selects' the sha256
digest as a component of its configuration so it is always
guaranteed to be present.

The -h command-line argument was added to all of the trust
orchestrators to allow the specification of the hash function.

In the current implementation, the external trust orchestrators
will only handle hash functions with a digest length of 32 bytes.
Specifying a hash function with an alternate digest length will
cause the model to fail.

The internal TMA implementation will handle all hash functions.