Implement JSON to quixote encoding conversion utility.
In order to facilitate working with the output from the TSEM
kernel LSM that is JSON encoded a utility was introducted that
allows the conversion of a security event description stream from
JSON encoding to 'quixote' encoding that is used by the current
trust orchestrators.

The utility takes a single arguement, -i, that specifies that
file that contains the JSON encoded security event descriptions
that are to be converted.  With no argument the standard input
stream is used.