Convert from hash encoded pathname to standard path name.
The Linux TSEM driver was converted to output a security event
description that has a file{} field with the standard pathname of
the file encoded as a 'path' key.  The 'name_length' key, that
gave the length of the filename was dropped, as was the 'name'
key that had held the hexadecimally encoded digest of the
filename.

The 'filename' key in the event{} field was dropped.  The notion
of surfacing the filename there wasn't consistent with the model
of the security event type description holding all of the
information needed to support the computation of the CELL
identity.  In addition, many of the events didn't have a filename
associated with them that necessitated encoding the filename as
'none'.

The CELL->_measure_file internal method was modified to compute
the hash of the filename so that the previous measurement values
would be reproduced.

This change allows a security model access to the file so that
future implementations can allow functionality such as treating
all files in a specific directory as pseudonyms or having a
common filename.