Update internal orchestrator for new count/points export.
The Linux TSEM driver's implementation of the deterministic
trusted modeling agent was enhanced to include the ability to
support the number of times a specific security event coefficient
was generated.

These counts are now specific to whether or not the events being
tracked are trajectory events or forensic events.  The
coefficients of these events are now also partitioned in the same
manner.

The TSEM implementation exposes the model parameters in the
following pseudo-files:

/sys/kernel/security/tsem/trajectory
/sys/kernel/security/tsem/trajectory_count
/sys/kernel/security/tsem/trajectory_points

/sys/kernel/security/tsem/forensics
/sys/kernel/security/tsem/forensics_count
/sys/kernel/security/tsem/forensics_points

The quixote trust orchestrator was modified to use these new
pseudo-files to return the model parameters to the
quixote-console utility.

The quixote-console utility was modified to request the output of
these parameters.

The following new commands were added:

show counts
show forensics_counts
show forensics_points

Which output the trajectory counts, the forensic event counts and
the forensics points respectively.

The 'show points' command now outputs the trajectory points.

The quixote-console utility had the -C command-line argument
added in order to support a one-shot mode which output the
trajectory counts.  The forensics counts and forensics points are
only available as command-line requests.