Support key/value definitions for namespace arguments.
The TSEM driver was updated to expect key=value arguments for the
parameters needed to create a security modeling domain.  This
replaces the former model of using fixed positional parameters.

The use of key=value pairs was implemented in order to provide
better for maintaining a userspace ABI.

One of the side effects of this is the ability to set default
parameters for a namespace.

One of the parameters support by this new ABI format is the size
of the atomic magazine structure caches that are to be
implemented for a modeling domain.  These atomic caches provide
the ability to model security events that are called in atomic
context.

The -n command-line argument was introduced to all of the trust
orchestrators to support providing a non-default magazine size.

The TSEMcontrol->create_ns method was modified to encode
parameters for the namespace creation call in key=value format.
The method was also modified so that if values of zero or NULL
are provided as arguments for the namespace parameters, the
argument is not encoded and the TSEM driver is allowed to use its
default value.

The TSEMcontrol->create_ns method was also modified to accept an
additional argument that allows the size of the magazine cache to
be specified.